# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
  models.AdminLoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.AdminReq:
    properties:
      name:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      username:
        type: string
    type: object
  models.AdminRes:
    properties:
      access_token:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      username:
        type: string
    type: object
  models.ClientLoginReq:
    properties:
      first_name:
        type: string
      phone_number:
        type: string
    type: object
  models.ClientUpdateReq:
    properties:
      age:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  models.ClientsList:
    properties:
      clients:
        items:
          $ref: '#/definitions/models.ClientsResponse'
        type: array
    type: object
  models.ClientsRequest:
    properties:
      age:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  models.ClientsResponse:
    properties:
      access_token:
        type: string
      age:
        type: integer
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is Clinic server api
  termsOfService: 1 term Traffic Light
  title: Clinic
  version: "1.0"
paths:
  /admin:
    post:
      consumes:
      - application/json
      description: Create Admin
      operationId: createadmin
      parameters:
      - description: CreateAdminInfo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AdminReq'
      produces:
      - application/json
      responses:
        "201":
          description: Admin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AdminRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Admin
      tags:
      - Admin
  /admin/{id}:
    get:
      consumes:
      - application/json
      description: Get Admin
      operationId: getadminbyid
      parameters:
      - description: admin_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Admin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AdminRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Admin
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: Login Admin
      operationId: loginadmin
      parameters:
      - description: AdminLogin
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.AdminLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Admin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AdminRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login Admin
      tags:
      - LOGIN
  /admins:
    get:
      consumes:
      - application/json
      description: Get all  Admin
      operationId: getadmin
      produces:
      - application/json
      responses:
        "200":
          description: Admin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AdminRes'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all Admin
      tags:
      - Admin
  /client:
    post:
      consumes:
      - application/json
      description: Create Client
      operationId: createclient
      parameters:
      - description: ClientCreate
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.ClientsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClientsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Client
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: PUT Clients
      operationId: putclients
      parameters:
      - description: UpdateClient
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ClientUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClientsList'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Clients
      tags:
      - Client
  /client/{id}:
    get:
      consumes:
      - application/json
      description: GET Client
      operationId: getclient
      parameters:
      - description: ClientID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClientsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: GET Client
      tags:
      - Client
  /client/delete/{id}:
    delete:
      consumes:
      - application/json
      description: PUT Clients
      operationId: deleteclients
      parameters:
      - description: clientID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ResponseMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Clients
      tags:
      - Client
  /client/login:
    post:
      consumes:
      - application/json
      description: Login Client
      operationId: loginclient
      parameters:
      - description: ClientLogin
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.ClientLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClientsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login Client
      tags:
      - LOGIN
  /clients:
    get:
      consumes:
      - application/json
      description: GET Clients
      operationId: getclients
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClientsList'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET Clients
      tags:
      - Client
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"